# failiverCheck/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
...
        "/system_calcs_to_components/": {
            "put": {
                "description": "Update component in system calc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Components To SystemCalc"
                ],
                "summary": "Update component in system calc",
                "parameters": [
                    {
                        "description": "Update schema",
                        "name": "create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateComponentToSystemCalcDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ComponentsToSystemCalcDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete component in system calc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Components To SystemCalc"
                ],
                "summary": "Delete component in system calc",
                "parameters": [
                    {
                        "description": "Delete schema",
                        "name": "create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ComponentToSystemCalcByIdDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ComponentInSystemCalcDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.ComponentToSystemCalcByIdDTO": {
            "type": "object",
            "required": [
                "component_id",
                "system_calculation_id"
            ],
            "properties": {
                "component_id": {
                    "type": "integer"
                },
                "system_calculation_id": {
                    "type": "integer"
                }
            }
        },
        "dto.ComponentsToSystemCalcDTO": {
            "type": "object",
            "properties": {
                "component": {
                    "$ref": "#/definitions/dto.ComponentInSystemCalcDTO"
                },
                "replication_count": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateComponentToSystemCalcDTO": {
            "type": "object",
            "properties": {
                "component_id": {
                    "type": "integer"
                },
                "replication_count": {
                    "type": "integer"
                },
                "system_calculation_id": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.Error": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "schemas.OKResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}